fix: registrytoken cmd notfound error (#3945)
The following PR fixes how registrytoken command should behave on
notfound errors.

Docker credential helpers must return a specific error on this scenario.

The PR also adds logs for what is being requested to the credential
helper.

An optimization is made to ensure the credential helper is not called by
okteto if set in the dockerconfig.

Finally, not-found api error is added to the translateApiErr function to
avoid the logline "Unrecognized API error".

Signed-off-by: Javier Provecho Fernández (Okteto) jpf@okteto.com

Signed-off-by: Javier Provecho Fernández (Okteto) jpf@okteto.com
(cherry picked from commit 686c5cbdbd698524602619ee836fabc8bcecd0e4)