fix: remove context init from rch install cmd
Context initialization is not needed for installing / uninstalling the
registry credential helper.

If initialized, environment needs to provide a valid auth and url.

These comands are meant to be executed in two places:
- remote dockerfile spec
- installer job init containers

The later is executed directly without the pipeline installer wrapper,
responsible for injecting OKTETO_TOKEN=$(POD_NAME).

Without this fix, the environment variable definition above will be at
plain sight in the pod spec of an installer job.

I've tested this by stashing it into a temporal branch from
release-2.19, building a custom okteto cli container image and fetching
it from a custom pipeline installer container image configured in my dev
environment. (screenshot in the first comment of the PR)

Signed-off-by: Javier Provecho Fernández (Okteto) jpf@okteto.com