feat: Add extractImageTagFromPullAccessDeniedError function
This commit adds a new function `extractImageTagFromPullAccessDeniedError` to the `errors.go` file. This function extracts the image tag from the error message when a pull access is denied. It uses a regular expression to find the image tag and returns it. This function is used in the `getErrorMessage` function to provide a more informative error message when pulling an inaccessible or non-existent image.

Co-authored-by: Nacho Fuertes <nacho@okteto.com>
Signed-off-by: Javier Lopez <javier@okteto.com>